jwt for each contact route

+++ controllers/contactController.js
@@ -9,16 +9,16 @@ user can access them ,
 */
 //@desc Get all contacts
 //@route GET/api/contacts
-//@access public
+//@access private
 const getContacts = asyncHandler(async (req, res) => {
-    const contacts = await Contact.find();
+    const contacts = await Contact.find({user_id: req.user.id});
     res.status(200).json(contacts);
 });
 
 
 //@desc Create new contact
 //@route POST/api/contacts
-//@access public
+//@access private
 const createContact = asyncHandler(async (req, res) => {
     console.log("the req body is : " , req.body);
 
@@ -35,6 +35,7 @@ const createContact = asyncHandler(async (req, res) => {
         name, 
         email, 
         phone,
+        user_id: req.user.id,
 
     });
     res.status(201).json(contact);
@@ -43,7 +44,7 @@ const createContact = asyncHandler(async (req, res) => {
 
 //@desc Get specific contact
 //@route GET/api/contacts/id
-//@access public
+//@access private
 const getContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
     if(!contact){
@@ -56,7 +57,7 @@ const getContact = asyncHandler(async (req, res) => {
 
 //@desc update contact
 //@route PUT/api/contacts/id
-//@access public
+//@access private
 const updateContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
     if(!contact){
@@ -64,6 +65,10 @@ const updateContact = asyncHandler(async (req, res) => {
         throw new Error("Contact not found");
     }
     
+    if(contact.user_id != req.user_id){
+        res.status(403);
+        throw new Error("This user does not have permission to edit this contact");
+    }
     const updatedContact = await Contact.findByIdAndUpdate(
         req.params.id, 
         req.body,
@@ -76,7 +81,7 @@ const updateContact = asyncHandler(async (req, res) => {
 
 //@desc delete contact
 //@route DELETE/api/contacts/id
-//@access public
+//@access private
 const deleteContact = asyncHandler(async (req, res) => {
     // Ensure to await the findById method to get the contact
     const contact = await Contact.findById(req.params.id);
@@ -86,6 +91,11 @@ const deleteContact = asyncHandler(async (req, res) => {
         throw new Error("Contact not found");
     }
 
+    if(contact.user_id != req.user_id){
+        res.status(403);
+        throw new Error("This user does not have permission to edit this contact");
+    }
+    
     // Use the remove method on the document instance (contact)
     await contact.deleteOne();
